refactor: update module imports to use index files (#33, #34)
