+++ src/modules/auth/auth.routes.js
@@ -1,18 +1,11 @@
-import {
-	register,
-	login,
-	googleAuth,
-	googleExtract,
-	sendCode,
-	linkGoogle,
-} from "./auth.service.js";
+import { authService } from "./index.js";
 import { checkAuth } from "../../shared/middleware/checkAuth.js";
 
 export default async function authRoutes(fastify) {
-	fastify.post("/register", register);
-	fastify.post("/login", login);
-	fastify.post("/google", googleAuth);
-	fastify.post("/google-extract", googleExtract);
-	fastify.post("/send-code", sendCode);
-	fastify.post("/link-google", { preHandler: checkAuth }, linkGoogle);
+	fastify.post("/register", authService.register);
+	fastify.post("/login", authService.login);
+	fastify.post("/google", authService.googleAuth);
+	fastify.post("/google-extract", authService.googleExtract);
+	fastify.post("/send-code", authService.sendCode);
+	fastify.post("/link-google", { preHandler: checkAuth }, authService.linkGoogle);
 }
