+++ src/modules/auth/auth.service.js
@@ -1,7 +1,7 @@
 import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
-import User from "../users/user.model.js";
-import TempCode from "./temp-code.model.js";
+import { User } from "../users/index.js";
+import { TempCode } from "./index.js";
 import { generateNickname } from "../../shared/utils/nicknameGen.js";
 import { verifyGoogleToken } from "../../infrastructure/google/googleClient.js";
 import { sendVerificationEmail } from "../../infrastructure/email/emailService.js";
