+++ src/modules/auth/index.js
@@ -1,0 +1,3 @@
+export { default as authRoutes } from './auth.routes.js';
+export * as authService from './auth.service.js';
+export { default as TempCode } from './temp-code.model.js';
