+++ src/modules/quizzes/index.js
@@ -1,0 +1,3 @@
+export { default as quizRoutes } from './quiz.routes.js';
+export * as quizService from './quiz.service.js';
+export { default as Quiz } from './quiz.model.js';
