+++ src/modules/quizzes/quiz.routes.js
@@ -1,23 +1,17 @@
-import {
-	getAllQuizzes,
-	createQuiz,
-	getQuizById,
-	updateQuiz,
-	deleteQuiz,
-} from "./quiz.service.js";
+import { quizService } from "./index.js";
 import { checkAuth } from "../../shared/middleware/checkAuth.js";
 
 export default async function quizRoutes(fastify) {
 	// public routes
-	fastify.get("/", getAllQuizzes);
-	fastify.get("/:id", getQuizById);
+	fastify.get("/", quizService.getAllQuizzes);
+	fastify.get("/:id", quizService.getQuizById);
 
 	// protected routes
 	fastify.register(async function (protectedRoutes) {
 		protectedRoutes.addHook("preHandler", checkAuth);
 
-		protectedRoutes.post("/", createQuiz);
-		protectedRoutes.put("/:id", updateQuiz);
-		protectedRoutes.delete("/:id", deleteQuiz);
+		protectedRoutes.post("/", quizService.createQuiz);
+		protectedRoutes.put("/:id", quizService.updateQuiz);
+		protectedRoutes.delete("/:id", quizService.deleteQuiz);
 	});
 }
