+++ src/modules/results/index.js
@@ -1,0 +1,3 @@
+export { default as resultRoutes } from './result.routes.js';
+export * as resultService from './result.service.js';
+export { default as Result } from './result.model.js';
