+++ src/modules/results/result.routes.js
@@ -1,10 +1,10 @@
-import { getUserResults, saveResult, getResultById } from "./result.service.js";
+import { resultService } from "./index.js";
 import { checkAuth } from "../../shared/middleware/checkAuth.js";
 
 export default async function resultRoutes(fastify) {
 	fastify.addHook("preHandler", checkAuth);
 
-	fastify.get("/", getUserResults);
-	fastify.get("/:id", getResultById);
-	fastify.post("/", saveResult);
+	fastify.get("/", resultService.getUserResults);
+	fastify.get("/:id", resultService.getResultById);
+	fastify.post("/", resultService.saveResult);
 }
