+++ src/modules/results/result.service.js
@@ -1,5 +1,5 @@
-import Result from "./result.model.js";
-import Quiz from "../quizzes/quiz.model.js";
+import { Result } from "./index.js";
+import { Quiz } from "../quizzes/index.js";
 
 // User results getting logic
 export const getUserResults = async (request, reply) => {
