+++ src/modules/users/index.js
@@ -1,0 +1,3 @@
+export { default as userRoutes } from './user.routes.js';
+export * as userService from './user.service.js';
+export { default as User } from './user.model.js';
