+++ src/modules/users/user.routes.js
@@ -1,24 +1,17 @@
-import {
-	getUser,
-	changePassword,
-	updateProfile,
-	deleteAccount,
-	getNicknameArray,
-	getUserById,
-} from "./user.service.js";
+import { userService } from "./index.js";
 import { checkAuth } from "../../shared/middleware/checkAuth.js";
 
 export default async function userRoutes(fastify) {
 	// public routes
-	fastify.get("/:id", getUserById);
+	fastify.get("/:id", userService.getUserById);
 
 	// protected routes
 	fastify.register(async function (protectedRoutes) {
 		protectedRoutes.addHook("preHandler", checkAuth);
-		protectedRoutes.get("/", getUser);
-		protectedRoutes.put("/update", updateProfile);
-		protectedRoutes.post("/password", changePassword);
-		protectedRoutes.delete("/delete", deleteAccount);
-		protectedRoutes.get("/nickname", getNicknameArray);
+		protectedRoutes.get("/", userService.getUser);
+		protectedRoutes.put("/update", userService.updateProfile);
+		protectedRoutes.post("/password", userService.changePassword);
+		protectedRoutes.delete("/delete", userService.deleteAccount);
+		protectedRoutes.get("/nickname", userService.getNicknameArray);
 	});
 }
